Move all tests to root __tests__
