+++ tsconfig.esm.json
@@ -6,14 +6,12 @@
     "outDir": "./esm"
   },
   "exclude": [
+    "__tests__",
     ".github",
     "coverage",
     "node_modules",
     "tests",
     "types",
-    "__tests__",
-    "**/__tests__",
-    "*.spec.ts",
     "utils"
   ]
 }
